// packages/diagram/src/likec4diagram/types.d.ts
export declare namespace Types {
// ...
    type SequenceActorNodeData = Simplify<LeafNodeData & {
        navigateTo: ViewId | null;
        modelFqn: Fqn | null;
        ports: Array<SequenceActorNodePort>;
        viewHeight: number;
    }>;
    type SequenceParallelAreaData = Simplify<LeafNodeData & {
        parallelPrefix: string;
    }>;
// ...
    type ViewGroupNodeData = Simplify<CompoundNodeData & {
        isViewGroup: true;
    }>;
    type ElementNode = BaseNode<ElementNodeData, 'element'>;
    type DeploymentElementNode = BaseNode<DeploymentElementNodeData, 'deployment'>;
    type SequenceActorNode = BaseNode<SequenceActorNodeData, 'seq-actor'>;
    type SequenceParallelArea = BaseNode<SequenceParallelAreaData, 'seq-parallel'>;
    type CompoundElementNode = BaseNode<CompoundElementNodeData, 'compound-element'>;
    type CompoundDeploymentNode = BaseNode<CompoundDeploymentNodeData, 'compound-deployment'>;
    type ViewGroupNode = BaseNode<ViewGroupNodeData, 'view-group'>;
// ...
    type SequenceStepEdgeData = Simplify<BaseEdgeData & NonOptional<Pick<DiagramEdge, 'id' | 'label' | 'technology' | 'points' | 'dir' | 'color' | 'line' | 'head' | 'tail' | 'navigateTo' | 'astPath' | 'drifts'>> & {
        notes: MarkdownOrString | null;
        labelXY: XYPosition | null;
        labelBBox: BBox;
        controlPoints: XYPosition[] | undefined | null;
    }>;
    type RelationshipEdge = BaseEdge<RelationshipEdgeData, 'relationship'>;
    type SequenceStepEdge = BaseEdge<SequenceStepEdgeData, 'seq-step'>;
// ...
}