// packages/diagram/src/likec4diagram/xyflow-sequence/const.d.ts
export declare const ACTOR_GAP = 60;
export declare const COLUMN_GAP = 32;
export declare const MIN_ROW_HEIGHT = 80;
export declare const PORT_HEIGHT = 32;
export declare const STEP_LABEL_MARGIN = 50;
export declare const FIRST_STEP_OFFSET = 30;
export declare const CONTINUOUS_OFFSET = 22;
export declare const SeqZIndex: {
    compound: number;
    parallel: number;
    actor: number;
    step: number;
};
export declare const SeqParallelAreaColor: {
    default: string;
    active: string;
};